Introduce OpenPolicyAgentClient trait
This commit introduces a trait for the OPA Client.
The motivation for this is to enable different implementations in
addition to the existing http client.

Fixes: #4
