+++ Cargo.toml
@@ -10,6 +10,7 @@ reqwest = { version = "0.11.12", features = [ "json"] }
 url = "2.3.1"
 serde = { version = "1.0.145", features = [ "derive" ] }
 serde_json = "1.0.85"
+async-trait = "0.1.58"
 
 [dev-dependencies]
 tokio = { version = "1", features = [ "macros" ]  }
